pyproject.toml
[project]
name = "computer-vision-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "joblib>=1.5.3",
    "mediapipe>=0.10.32",
    "opencv-python>=4.13.0.92",
    "scikit-learn>=1.8.0",
    "streamlit>=1.42.0",
    "streamlit-webrtc>=0.47.10",
    "av>=14.0.1",
    "python-fasthtml>=0.12.50",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
]
